backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dateutil==2.8.2
drf-yasg==1.21.7
Faker==18.3.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0